Moving useNetwork to context api
