+++ src/contexts/index.ts
@@ -2,6 +2,7 @@ import ThemeProvider, { ThemeContext } from './themeContext/ThemeContext';
 import AmplitudeProvider, { AmplitudeContext } from './AmplitudeContext/AmplitudeContext';
 import BiconomyProvider, { BiconomyContext } from './BiconomyContext/BiconomyContext';
 import LanguageProvider, { LanguageContext } from './LanguageContext/LanguageContext';
+import NetworkProvider, { NetworkContext } from './NetworkContext/NetworkContext';
 import TransactionsProvider, { TransactionsContext } from './TransactionsContext/TransactionsContext';
 
 export {
@@ -10,6 +11,7 @@ export {
 	LanguageContext,
 	BiconomyContext,
 	TransactionsContext,
+	NetworkContext,
 	TransactionsProvider
 };
 
@@ -17,5 +19,6 @@ export const providers = [
 	ThemeProvider,
 	AmplitudeProvider,
 	BiconomyProvider,
+	NetworkProvider,
 	LanguageProvider
 ];
